Refactor map components to eliminate duplicate code and fix production database issues
- Create shared BaseMap component for common map initialization, 3D viewport, and controls
- Refactor FlightMap, FlightsMap, PilotMap, and SiteMap to use BaseMap with specific callbacks
- Eliminate ~1000+ lines of duplicate code while preserving all functionality
- Add explicit PostgreSQL type casting (::integer) to fix production database type errors
- Convert URL parameters from string to number for pilot_id fields
- Update infrastructure for proper Mapbox token deployment and environment handling
- Fix Taskfile.yml to update terraform.tfvars line-by-line rather than overwriting

+++ site/src/data/database/Sites.ts
@@ -49,7 +49,7 @@ export namespace Sites {
                                    count(1)   as flights
                             from flights as f
                                      left join sites as s on s.ffvl_sid = f.takeoff_id
-                            where f.pilot_id = $1
+                            where f.pilot_id = $1::integer
                             group by f.takeoff_id, s.ffvl_sid)
 
                 select ss.flights as flights, to_json(s) as site
@@ -63,7 +63,7 @@ export namespace Sites {
                                    count(1)   as flights
                             from flights as f
                                      left join sites as s on s.ffvl_sid = f.landing_id
-                            where f.pilot_id = $1
+                            where f.pilot_id = $1::integer
                             group by f.landing_id, s.ffvl_sid)
 
                 select ss.flights as flights,
